fix(sas): 添加新闻评论

+++ src/main/java/com/food/sas/data/dto/FocalPictureDTO.java
@@ -23,5 +23,5 @@ public class FocalPictureDTO implements Serializable {
 
     private Date createDate;
 
-    private Integer creator;
+    private Long creator;
 }
